Move settings to respective paths as listed in TODO.MD
